pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptparse"
version = "0.0.1"
authors = [{ name = "Hao Zheng", email = "wszh712811@gmail.com" }]
description = "PPTParse, a tool for parsing and building PowerPoint presentations."
requires-python = ">=3.10"
license = "MIT"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3",
    "Topic :: Office/Business :: Office Suites",
]
keywords = [
    "presentation",
    "powerpoint",
    "ppt",
    "document-processing",
]
dependencies = [
    "python-pptx @ git+https://github.com/Force1ess/python-pptx@219513d7d81a61961fc541578c1857d08b43aa2a",
    "tenacity",
    "dacite",
]

[project.urls]
"Homepage" = "https://github.com/Force1ess/PPTParse"
"Bug Tracker" = "https://github.com/Force1ess/PPTParse/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["pptparse*"]
